rokid_ai_agent/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.0.0
dashscope>=1.14.0
faiss-cpu>=1.7.0
httpx>=0.25.0
pydantic>=2.0.0
sse-starlette>=1.6.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
